Problem 96 with scala.
One commit, so many evenings and attempts :)

TODO: make concurrent with Future and handle case where multiple
direct updates are available which can currently lead to issues so only
one update is now done per iteration.

----------

Congratulations, the answer you gave to problem 96 is correct.

You are the 16381st person to have solved this problem.

This problem had a difficulty rating of 25%. The highest difficulty rating you have solved so far is 30%.